requirements.txt
binance-connector==3.7.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
